rename core files
