+++ README.md
@@ -29,7 +29,7 @@ For this wrapper to run, you need an authenticated Zesty.io user APP_SID and a I
 const instanceZUID = '8-xyzxyz-xyz'
 const userAppSID = 'xxxxxxxxxx'
 
-ZestyAPI = new ZestyioRestFetchWrapper(instanceZUID,userAppSID)
+ZestyAPI = new $.ZestyioRestFetchWrapper(instanceZUID,userAppSID)
 
 console.log(ZestyAPI.getModels())
